Add cpu mode support
This adds a new config option in the libvirt section
named cpu_mode. If the virt type is KVM this will default
to cpu-passthrough and replace the hardcoded value that is
currently there today and thus keeping backward compatibility.

This allows to boot RHEL 9/CentOS Stream 9 using QEMU without
KVM by setting cpu_mode=host-model since it does not support
the default qemu64/kvm64 cpu models that is default.